pyproject.toml
[tool.poetry]
name = "djangocon-eu-ddd"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
strawberry-graphql = "^0.64.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-django = "^4.3.0"
flake8 = "^3.9.2"
mypy = "^0.812"
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
